+++ src/shared/assets/icons/sidbarIcon/2_create/create2.png
