+++ src/shared/assets/icons/sidbarIcon/5_search/serch2.png
