+++ src/shared/assets/icons/sidbarIcon/6_statistics/statistics2.png
