+++ src/shared/assets/icons/sidbarIcon/7_favorites/favorites2.png
