+++ src/shared/assets/icons/sidbarIcon/8_log_out/account2.png
