+++ src/widgets/sidebar/ui/Sidebar.tsx
@@ -1,5 +1,6 @@
 import React from 'react'
 
+import { getSidebarEl } from '@/shared/assets/icons/sidbarIcon/index'
 import { translate } from '@/shared/lib/langSwitcher'
 import { SidebarPng } from '@/widgets/sidebar/ui/sidebar-png/SidebarPng'
 import clsx from 'clsx'
@@ -7,8 +8,6 @@ import Link from 'next/link'
 import { useRouter } from 'next/router'
 
 import s from './sidebar.module.scss'
-
-import { getSidebarEl } from '../pmg/index'
 
 type TypeProps = {
   className?: string
@@ -24,7 +23,7 @@ export const Sidebar: React.FC = ({ className, pngSize }: TypeProps) => {
     <aside className={s.border_in}>
       <div className={clsx(clsx(s.space, s.space_top), [className])}>
         {top.map((el, i) => (
-          <Link className={''} href={el.path} key={i} passHref>
+          <Link href={el.path} key={i} passHref>
             <SidebarPng className={s.link} name={el.name} png={el.png} size={pngSize || 24} />
           </Link>
         ))}
@@ -32,7 +31,7 @@ export const Sidebar: React.FC = ({ className, pngSize }: TypeProps) => {
 
       <div className={clsx(clsx(s.space, s.space_top_bottom), [className])}>
         {middle.map((el, i) => (
-          <Link className={''} href={el.path} key={i} passHref>
+          <Link href={el.path} key={i} passHref>
             <SidebarPng className={s.link} name={el.name} png={el.png} size={pngSize || 24} />
           </Link>
         ))}
@@ -40,7 +39,7 @@ export const Sidebar: React.FC = ({ className, pngSize }: TypeProps) => {
 
       <div className={clsx(clsx(s.space, s.space_bottom), [className])}>
         {bottom.map((el, i) => (
-          <Link className={''} href={el.path} key={i} passHref>
+          <Link href={el.path} key={i} passHref>
             <SidebarPng className={s.link} name={el.name} png={el.png} size={pngSize || 24} />
           </Link>
         ))}
